% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotocasa.R
\name{fotocasa}
\alias{fotocasa}
\title{Scrap Fotocasa website.}
\usage{
fotocasa(url, ruta = "~/fotocasa.csv")
}
\arguments{
\item{url}{A Fotocasa website url that links to the area you want to scrap, e.g. 'https://www.fotocasa.es/es/alquiler/casas/madrid-capital/todas-las-zonas/l'.}

\item{ruta}{A valid path in your computer where you want to create the csv file.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps Fotocasa (a spanish real estate website) and downloads all the rent ads in the given province, city, disctrict or neighborhood.
}
