% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distrito_.R
\name{get_distrito_}
\alias{get_distrito_}
\title{Scrap idealista website.}
\usage{
get_distrito_(url, ads, ruta = "~/idealisto_distrito.csv", silent = FALSE)
}
\arguments{
\item{url}{character. An idealista website url that links to the city district you want to scrap, e.g. 'https://www.idealista.com/alquiler-viviendas/madrid/barajas/'.}

\item{ads}{character. Specify if the url links to rent ads or for sale ads. The argument accepts the following strings: "rent" or "sale".}

\item{ruta}{character. A valid path in your computer where you want to create the csv file.}

\item{silent}{logical. If TRUE it will print less messages. Useful if you want to schedule with cron and want a cleaner log file. The default is FALSE.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps idealista (a spanish real estate website) and downloads all the rent ads of a given city district.
}
