% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provincia.R
\name{get_provincia}
\alias{get_provincia}
\title{Scrap idealista website.}
\usage{
get_provincia(url, ruta = "~/idealisto_prov.csv", silent = FALSE)
}
\arguments{
\item{url}{character. An idealista website url that links to the province you want to scrap, e.g. 'https://www.idealista.com/alquiler-viviendas/zaragoza-provincia/'.}

\item{ruta}{character. A valid path in your computer where you want to create the csv file.}

\item{silent}{logical. If TRUE it will print a less messages. Useful if you want to schedule with cron and want a cleaner log file. The default is FALSE.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps idealista (a spanish real estate website) and downloads all the rent ads of a given province
}
