% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealisto_silent.R
\name{idealisto_silent}
\alias{idealisto_silent}
\title{Scrap idealista website.}
\usage{
idealisto_silent(url, area, ruta = "~/idealisto_alter.csv")
}
\arguments{
\item{url}{An idealista website url that links to the area you want to scrap, e.g. 'https://www.idealista.com/alquiler-viviendas/madrid/arganzuela/'.}

\item{area}{The type of area you want to scrap. It can take these values: 'Provincia', 'Ciudad', 'Distrito' or 'Barrio'.}

\item{ruta}{A valid path in your computer where you want to create the csv file.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps idealista (a spanish real estate website) and downloads all the rent ads in the given province, city, disctrict or neighborhood.
}
\details{
This function is a silent version of idealisto_alter. This means that it will print a less messages. Useful if you want to schedule with cron and want a cleaner log file.
}
