% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealisto_venta.R
\name{idealisto_venta}
\alias{idealisto_venta}
\title{Scrap idealista website.}
\usage{
idealisto_venta(url, area, ruta)
}
\arguments{
\item{url}{An idealisto website url that links to the area you want to scrap, e.g. 'https://www.idealista.com/alquiler-viviendas/madrid/arganzuela/'.}

\item{area}{The type of area you want to scrap. It can take these values: 'Provincia', 'Ciudad', 'Distrito' or 'Barrio'.}

\item{ruta}{A valid path in your computer where you want to create the csv file.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps idealista (a spanish real estate website) and downloads all the 'for sale' ads in the given province, city, district or neighborhood.
}
