% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigencia.R
\name{vigencia}
\alias{vigencia}
\title{Enrich the csv file generated by idealisto function.}
\usage{
vigencia(ruta)
}
\arguments{
\item{ruta}{A valid path in your computer that leads to an idealisto created csv file.}
}
\value{
It returns a csv file with more variables. It will be saved in the root (~/).
}
\description{
This function enriches the csv file generated by idealisto with some other variables related with the current situation of the downloaded ads. It's useful to know if the ads are still online days or weeks later or have been removed. If the ad has been removed, vigencia function will tell how many days ago and if the ad is still online it will tell if they have a new price.
}
