% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Idealisto_alter.R
\name{idealisto_alter}
\alias{idealisto_alter}
\title{Scrap idealista website.}
\usage{
idealisto_alter(url, area, ruta = "~/idealisto_alter.csv")
}
\arguments{
\item{url}{An idealisto website url that links to the area you want to scrap, e.g. 'https://www.idealista.com/alquiler-viviendas/madrid/arganzuela/'.}

\item{area}{The type of area you want to scrap. It can take these values: 'Provincia', 'Ciudad', 'Distrito' or 'Barrio'.}

\item{ruta}{A valid path in your computer where you want to create the csv file.}
}
\value{
It returns a csv in the specified path
}
\description{
This function scraps spanish idealista (a real estate website) and downloads all the rent ads in the given province, city, disctrict or neighborhood.
}
