% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigencia.R
\name{vigencia}
\alias{vigencia}
\title{Enrich the csv file generated by idealisto function.}
\usage{
vigencia(df, ruta = NULL)
}
\arguments{
\item{df}{A data frame object generated by idealisto function and imported later to the enviroment.}

\item{ruta}{A valid path in your computer where you want to create the csv file. If none given it will generate the csv in the working directory.}
}
\value{
It returns a data.frame object with the new columns.
}
\description{
This function enrichs the csv file generated by idealisto with some other variables related with the current situation of the ads downloaded. It's useful to know if the ads are still online days later or have been removed. If the ad has been removed, vigencia function will tell how many days ago and if the ad is still online it will tell if they have a new price. First you have to import the csv file and then call the function on the data frame object.
}
